+++ remix/server/routes/api/[...].ts
@@ -28,6 +28,8 @@ export const routeModules: Record<string, () => Promise<RouteModule>> = {
   'v1/admin/users': () => import('../../../app/routes/api/v1/admin/users/_users'),
   'v1/admin/users/overview': () => import('../../../app/routes/api/v1/admin/users/overview/_overview'),
   'v1/admin/users/public-uploads': () => import('../../../app/routes/api/v1/admin/users/public-uploads/_public-uploads'),
+  'v1/peers': () => import('../../../app/routes/api/v1/peers/_peers'),
+  'v1/peers/sync': () => import('../../../app/routes/api/v1/peers/sync/_sync'),
   'v1/integrations/github/webhook': () => import('../../../app/routes/api/v1/integrations/github/webhook/_webhook'),
   'v1/integrations/ci/route': () => import('../../../app/routes/api/v1/integrations/ci/route/_route'),
   'v1/integrations/vercel/webhook': () => import('../../../app/routes/api/v1/integrations/vercel/webhook/_webhook'),